Create Literally
Reading from a text ﬁle, you are to write a program that counts the number of times each 
individual letter of the alphabet occurs in that ﬁle, counting upper and lower case letters as the 
same. This tally is to be stored in an array called tally, so tally[0] holds the number of As, 
tally[1] holds the number of Bs, and so on. You must also calculate the percentage of use for 
each letter.
OUTPUT: 1) The ﬁle exactly as it is stored.
2) The tally of each letter.
3) The percentage, rounded to tenths, of use of each letter in relation to the 
 total number of letters in the ﬁle. These need not be stored.
NOTE: You are expected to do this without using nested ifs or a switch for the tallying.
 The cool part of this program is how the tallying can be generalized.
 The class is to have as private data the array and the total number of letters.
Make a text ﬁle with the following sentence: 
The quick brown fox jumps over the lazy dog.
Run the program ﬁrst with that ﬁle, then with the paragraph.dat ﬁle that will be provided. 
The user is to enter the ﬁle name.